requirements.txt
black==21.12b0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
commonmark==0.9.1
docutils==0.18.1
feedgenerator==2.0.0
Flask==2.1.1
flask-ldap3-login==0.9.17
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
ldap3==2.9.1
Markdown==3.3.6
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pyasn1==0.4.8
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==3.5.3
requests==2.27.1
rich==12.3.0
six==1.16.0
SQLAlchemy==1.4.35
tomli==1.2.3
typing_extensions==4.0.1
Unidecode==1.3.4
urllib3==1.26.9
Werkzeug==2.1.1
WTForms==3.0.1
